add a BIOS-based disc loader
The BIOS exposes three calls that init, read, and boot a CD.
All three have problems that mean Pseudo can't use them as is:
init starts a disc auth, read checks a disc auth before doing anything,
and boot checks the region.

In this loader, the first two functions are emulated. The third is
interesting: the read function reads sectors into the CDB - not into the
SH-2 memory. So we can just use the Put Sector Data functions to put
what we like in to be booted: in this case, the disc's real IP, but with
the region patched, and any Pseudo header overwritten with a Sega one.

The boot function is kind enough to do lots of header checking for us,
and returns meaningful error codes. It also won't clobber any memory
unless it's not coming back (aside from the IP load at 6002000), so we
don't need to use cdload to relocate and boot.

This patch leaves the original boot methods as a fallback; pressing A at
the error message will continue.